Reorganized home page code into separate module
